PBAP: Clean raw contacts before each new downloading
In the current implementation, after the PBAP connection is established,
the contacts will be downloaded and stored in the database. Later, if PBAP
connection is released, the stored contacts will be deleted by removing the
related account. However, in a special test case, if the user restarts the
system after the contacts are stored in the database, when system restarts
and the PBAP reconnection is completed, duplicate contacts of the same
account can be observed in the contacts application.

This change is to provide a solution to delete the outdated contacts each
time before starting to download new contacts.

Change-Id: Ifc086fce670262e7d544dbba893eb85733af4228