HF_CLIENT: Set conference call in proper state
Issue found:
Supposing in the scenario that a conference call is in holding state and
an incoming call is in active state, at this moment, reenabling BT in
Settings. After HFP connection gets established again, the active incoming
call is merged into the conference call unexpectedly.

Root cause:
When the instance of HfpClientConference is created, the initial state is
set as ACTIVE by default. This triggers wrong operations in onConnectionAdded.

Resolution:
Set a proper state for the conference instance before calling mConference.addConnection.

Change-Id: I47ae42c373bafd6f6301ffa75aad8febfe6412ba