AVRCP: Set active player when add player to player list
Some media players may not be available and cannot be added into media players
list When AVRCP startup. When application bonds player with device the process
to set active player may not be completed. So set the player to active player
when the player is added into media player list.

Change-Id: Ibae322e62b333029a72a525a7eb9b77ea30e5175