ProfileService: Avoid duplicated start() for profile service
Issue found:
initNative() is invoked repeatedly which results in the failure
of profile connection request.

Root cause:
Profile service may be started repeatedly due to the implementation
of application. In current implementation of doStart() in ProfileService,
it does not check the flag mProfileStarted which indicates whether the
target profile service is started yet or not. Duplicated invocation of
start() results in mistaken behavior in JNI.

Resolution:
Add flag check in doStart().

Change-Id: I5da53331fb45a4d77b0fc3c2addcee4adcadf130