AVRCP_CT: Clear cover art cache when AVRCP is disconnected
Issue found:
Cover art of the previously connected device is displayed.

Root cause:
The reason is that the cover art cache for the previous AVRCP
connection is not cleared. In current implementation, Bluetooth
framework fetches cover art images from local cache instead of
peer AVRCP target device if the cover art handle of the image to
be fetched can be found in local cache.

Resolution:
Always clear the cover art cache when AVRCP is disconnected.

Change-Id: Iaaa0255b22fbd5e5f9d24478bd39a144b1f5c532