MAP_CLIENT: Catch IllegalArgumentException
Issue found:
Observe Bluetooth process get crashed after MAP connection is
established.

Root cause:
Sometimes if the received OBEX packet does not comply with the
parsing rules, it may cause the parser to throw the exception
'IllegalArgumentException'. If the Bluetooth process does not
catch this exception, the process will be killed unexpectedly.

Resolution:
Add code in Request.java to catch this exception.

Change-Id: If6ba9e0d1e587434a0e526664de5389327fa9a13