AVRCP_CT: Adjust processing order of messages
Issue found:
The cover art image is not displayed after the addressed
player is changed.

Root cause:
In current implementation of the CmdState 'GetPlayerListing',
MESSAGE_PROCESS_TRACK_CHANGED is processed at once while
MESSAGE_PROCESS_ADDRESSED_PLAYER_CHANGED is deferred to be
processed in the next CmdState. However, this may cause the
processing order be different with the actual receiving order
in BT stack. Therefore, it results in the cover art handle in
subsequent BIP_IMAGE_FETCHED message be invalid.

Resolution:
Change the processing of MESSAGE_PROCESS_TRACK_CHANGED in CmdState
'GetPlayerListing' from instant to deferred.

Change-Id: Iad550c449a5de99d013f372bd3f56e3bf521e008