A2DP: Fix sound is mixed when select 2 media players for 1 headset
Issue:
When connect a headset and select media player A, and then select
media player B for this headset, the 2 media player player sound
is mixed when hear on headset.

Root cause:
In A2dpService.getAudioZoneIndex() , when clear old media player
with the deivce, audio zone is not unbinded with old media player,
and led to same audio zone is used by 2 media player.

Fix:
Clear audio zone id with old media player when select a new player

Change-Id: Ieabe140373091e7540248b93bcebf3c2782e1ae9