Updating id in hfp client
 - Updating the id with number Hash Code to distinguish such that
avoiding race conditions between two calls.


Issue:
In some cases when there is an active client call on dut and we get
an incoming call, same time just before getting MT call user terminate
the MO call from companion.

Here the expectation is MO call should be terminated and right number
should be displayed for incoming call.


Due to race condition in Active call termination and MT call,
Companion is responding current calls where same id as

on dailer app phone number of outgoing call is observed as the details
of incoming call.
This is because when we have initiated the incoming call the id is 1.
when we terminate the active call and meanwhile received incoming call
the call id of new incoming call is also 1.
so when we query about call details it is considering as same active call
as the id is 1. and a new UUID is not created, direction is not changed
and only state is being changed which means not considered as a new
MT call.

Fix:
When query is done we can't rely on just id's as this is creating the
issue like above.
So we should identify the calls based on phone numbers and id's.
For this reason when query is done instead of id we updated
(id + number.hashCode() ) which generates a unique id for different calls.
Therefore considering each call different as their id's are different.

Change-Id: Ic18b8f066c6b871c4deed35346ca86f7eb59ebed

CRs-Fixed: 3683925
(cherry picked from commit 538b252e00691aef2a1a471bdfd47d954f48bcff)

Change-Id: Ia7ab678160cdf7be33cee03e03ae0f5bba2124c2