a2dp:  Set the active device to null after the A2DP profile is disconnected.
issue: While streaming music over Bluetooth, if the device is disconnected during streaming, a race condition occurs. As a result, ATH informs the system to set the active device to ‘slate’ before it receives the disconnection callback for the remote device.

fix: Set the active device to null after the A2DP profile is disconnected.

Change-Id: I8baa12e40df07c356d396177a78f99720006ee65
CRs-Fixed: 3907617