HFP: sending CLCC indicators to all the connected devices.
When 2 devices are connected and both of the remotes sends
CLCC request, the request has been added in the State machine
Queue. while responding to CLCC response for one of the remote,
we are unconditionally clearing all the messages in the queue.

Added changes to keep synchronized count on the requests received
and responses sent.

CRs-Fixed: 3969981
Change-Id: I81d40ee41d5155d271306e18b424602e6ca02f4d