HFP_Client: Fix screen flickering issue when merging a call to conference
Description:
Screen flickering is observed when merging a call into a conference
call.

Root Cause:
The call is added to the conference call directly, without waiting
for the update from the Audio Gateway. The subsequent call status
update indicates that the call is not immediately merged because
call merging takes time.

Solution:
Use onConnectionAdded instead of addConnection when to merge call
to conference call.

Change-Id: I723c2434c20544cd2672e3b3c74ad664b12076ea