hearing aid:  Broadcast Hearing Aid Device Active
Issue: The active hearing aid device is not being communicated to the telephony system. As a result, when a call is attempted, the hearing aid is not being selected for audio. It is essential to keep other modules informed about the active device for the hearing aid profile.

Fix: Broadcast the active hearing aid device status immediately after connection.

Change-Id: I74457301ed9bf1e768c8a542b68d40298c3138c6
CRs-Fixed: 4047415