HFP_Client: Fix exception in dialing a call
Issue Description:
BT exception when dialing a call

Root Cause:
In the constructor of HfpClientConnection, handleCallChanged
is called to process the call state change. If the state is
TERMINATED, mCurrentCall is set to null. Consequently, a null
pointer exception occurs when mCurrentCall is accessed in the
finishInitializing function.

Solution:
Don't call buildConnection if call is already in TERMINATED
state.

Change-Id: I2d01735aa1010a153e0c5c77d8f5548de6b5863d